requirements.txt
openai
langchain_openai
langchain
langchain_community
langchain_core 
python-dotenv
streamlit

langserve
fastapi
uvicorn
sse_starlette
pydantic==1.10.13